% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geocbi.R
\name{add_geocbi}
\alias{add_geocbi}
\title{Add the GeoCBI calculated from the burn grading data}
\usage{
add_geocbi(data)
}
\arguments{
\item{data}{A data.frame of burn grading data, collected with ODK Collect
from the form "Burn Grading 0.1" and downloaded with
\code{ruODK::odata_submission_get(wkt=T)}.}
}
\value{
The input data.frame with an extra variable, 'geocbi'.
}
\description{
Add the GeoCBI calculated from the burn grading data
}
