% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ozcbi.R
\name{add_ozcbi}
\alias{add_ozcbi}
\title{Add the OzCBI calculated from the burn grading data}
\usage{
add_ozcbi(burngrading)
}
\arguments{
\item{burngrading}{A data.frame of burn grading data,
collected with ODK Collect from the form "Burn Grading 0.1"
and downloaded with \code{ruODK::odata_submission_get(wkt=T)}.}
}
\value{
The input data.frame with an extra variable, 'ozcbi'.
}
\description{
Add the OzCBI calculated from the burn grading data
}
\examples{
data("burngrading01")
data_with_ozcbi <- add_ozcbi(burngrading01)
names(data_with_ozcbi)
}
\seealso{
Other ozcbi: 
\code{\link{calculate_ozcbi}()}
}
\concept{ozcbi}
