% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ozcbi_forest.R
\name{add_ozcbi_forest}
\alias{add_ozcbi_forest}
\title{Add the OzCBI calculated from the burn grading data}
\usage{
add_ozcbi_forest(burngrading)
}
\arguments{
\item{burngrading}{A data.frame of burn grading data,
collected with ODK Collect from the form "Burn Grading 0.1"
and downloaded with \code{ruODK::odata_submission_get(wkt=T)}.}
}
\value{
The input data.frame with an extra variable, 'ozcbi'.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
data("bgf")
data_with_ozcbi <- add_ozcbi_forest(bgf)
names(data_with_ozcbi)
}
\seealso{
Other ozcbi: 
\code{\link{calculate_ozcbi_forest}()}
}
\concept{ozcbi}
