% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bgv}
\alias{bgv}
\title{Burn Grading Verify form submission data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 70 columns.
}
\source{
Form submissions of form
\code{system.file("odk/BurnGradingVerify.xlsx", package = "rOzCBI")}
downloaded from ODK Central using \code{ruODK::odata_submission_get()}
(see \code{vignette("analysis")}).
}
\usage{
bgv
}
\description{
Burn Grading Verify form submission data
}
\seealso{
Other included: 
\code{\link{bgf}},
\code{\link{bgh}},
\code{\link{burngrading01}},
\code{\link{fms}}
}
\concept{included}
\keyword{datasets}
