% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burngrading01}
\alias{burngrading01}
\title{Burn Grading 0.1 form submission data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 56 columns.
}
\source{
Form submissions of form
\code{system.file("odk/archive/BurnGrading01.xml", package = "rOzCBI")}
downloaded from ODK Central using \code{ruODK::odata_submission_get()}
(see \code{vignette("analysis")}).
}
\usage{
burngrading01
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
Use data("bgf") instead.
}
\seealso{
Other included: 
\code{\link{bgf}},
\code{\link{bgh}},
\code{\link{bgv}},
\code{\link{fms}}
}
\concept{included}
\keyword{datasets}
