% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grading.R
\name{extract_grading}
\alias{extract_grading}
\title{Extract the grading from a string}
\usage{
extract_grading(grading_string)
}
\arguments{
\item{grading_string}{A string in the format <grading>_<label>.}
}
\value{
The grading as numeric value.
}
\description{
Extract the grading from a string
}
\examples{
testthat::expect_equal(parse_grading("0.0_test"), 0.0)
testthat::expect_equal(parse_grading("0.5_test"), 0.5)
testthat::expect_equal(parse_grading("1.0_test"), 1.0)
testthat::expect_equal(parse_grading("1.5_test"), 1.5)
testthat::expect_equal(parse_grading("2.5_test-with-long-label"), 2.5)
}
