% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fms}
\alias{fms}
\title{FMS verify severity form submission data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 70 columns.
}
\source{
Form submissions of form
\code{system.file("odk/FMSVerifySeverity.xml", package = "rOzCBI")}
downloaded from ODK Central using \code{ruODK::odata_submission_get()}
(see \code{vignette("analysis")}).
}
\usage{
fms
}
\description{
FMS verify severity form submission data
}
\seealso{
Other included: 
\code{\link{burngrading01}}
}
\concept{included}
\keyword{datasets}
