% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_burngrading.R
\name{map_burngrading}
\alias{map_burngrading}
\title{Map Burn Grading data (incl GeoCBI) with labels and popups}
\usage{
map_burngrading(data)
}
\arguments{
\item{data}{Data from form Burn Grading with GeoCBI.}
}
\value{
A leaflet map
}
\description{
Map Burn Grading data (incl GeoCBI) with labels and popups
}
\examples{
\dontrun{
data("burngrading01")
burngrading01 \%>\%
  add_geocbi() \%>\%
  map_burngrading()
}
}
\seealso{
Other helpers: 
\code{\link{extract_grading}()}
}
\concept{helpers}
