% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_burngrading.R
\name{reactable_burngrading}
\alias{reactable_burngrading}
\title{Display a reactable of Burn Grading data (incl GeoCBI) with details}
\usage{
reactable_burngrading(data)
}
\arguments{
\item{data}{Data from form Burn Grading with GeoCBI.}
}
\value{
An interactive reactable.
}
\description{
Display a reactable of Burn Grading data (incl GeoCBI) with details
}
\examples{
\dontrun{
data("burngrading01")
burngrading01 \%>\%
  add_geocbi() \%>\%
  reactable_burngrading()
}
}
\seealso{
Other helpers: 
\code{\link{extract_grading}()},
\code{\link{map_burngrading}()}
}
\concept{helpers}
