% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_ozcbi_forest.R
\name{reactable_ozcbi_forest}
\alias{reactable_ozcbi_forest}
\title{Display a reactable of Burn Grading Forest data (incl OzCBI) with details}
\usage{
reactable_ozcbi_forest(data)
}
\arguments{
\item{data}{Data from form Burn Grading Forest with OzCBI.}
}
\value{
An interactive reactable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
data("bgf")
bgf \%>\%
  add_ozcbi_forest() \%>\%
  reactable_ozcbi_forest()
}
}
\seealso{
Other helpers: 
\code{\link{extract_grading}()},
\code{\link{map_ozcbi_forest}()}
}
\concept{helpers}
