% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGeoCBI-package.R
\docType{package}
\name{rGeoCBI-package}
\alias{rGeoCBI}
\alias{rGeoCBI-package}
\title{rGeoCBI: A burn severity assessment toolkit}
\description{
rGeoCBI contains background, digital data capture, analysis, and
             visualisation for the calculation of GeoCBI, a composite burn index
             for the initial assessment of the short-term burn severity from
             remotely sensed data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dbca-wa.github.io/rGeoCBI}
  \item \url{https://github.com/dbca-wa/rGeoCBI}
  \item Report bugs at \url{https://github.com/dbca-wa/rGeoCBI/issues}
}

}
\author{
\strong{Maintainer}: Florian Mayer \email{Florian.Mayer@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-4269-4242}{ORCID})

Authors:
\itemize{
  \item Valerie Densmore \email{Valerie.Densmore@dbca.wa.gov.au}
}

Other contributors:
\itemize{
  \item DBCA [copyright holder, funder]
}

}
\keyword{internal}
