% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOzCBI-package.R
\docType{package}
\name{rOzCBI-package}
\alias{rOzCBI}
\alias{rOzCBI-package}
\title{rOzCBI: A Burn Severity Assessment Toolkit for Austral Vegetation Communities}
\description{
rOzCBI contains background, digital data capture, analysis, and
             visualisation for the calculation of OzCBI, a composite burn index
             for the initial assessment of the short-term burn severity from
             remotely sensed data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dbca-wa.github.io/rOzCBI}
  \item \url{https://github.com/dbca-wa/rOzCBI}
  \item Report bugs at \url{https://github.com/dbca-wa/rOzCBI/issues}
}

}
\author{
\strong{Maintainer}: Florian Mayer \email{Florian.Mayer@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-4269-4242}{ORCID})

Authors:
\itemize{
  \item Valerie Densmore \email{Valerie.Densmore@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-0121-8709}{ORCID})
}

Other contributors:
\itemize{
  \item DBCA [copyright holder, funder]
}

}
\keyword{internal}
