% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burngrading01}
\alias{burngrading01}
\title{Burn Grading 0.1 form submission data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 56 columns.}
\source{
Form submissions of form
\code{system.file("odk/BurnGrading01.xml", package = "rOzCBI")}
downloaded from ODK Central using \code{ruODK::odata_submission_get()}
(see \code{vignette("analysis")}).
}
\usage{
burngrading01
}
\description{
Burn Grading 0.1 form submission data
}
\concept{included}
\keyword{datasets}
