% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grading.R
\name{extract_grading}
\alias{extract_grading}
\title{Extract the grading from a Burn Grading field value}
\usage{
extract_grading(grading_string)
}
\arguments{
\item{grading_string}{A string in the format \verb{<grading>_<label>}.
These strings are field values of the form "Burn Grading".
E.g. "0.5_test-label", "2.0_other-label".}
}
\value{
The grading as numeric value, e.g. 0.5, 2.0.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
testthat::expect_equal(extract_grading("0.0_test"), 0.0)
testthat::expect_equal(extract_grading("0.5_test"), 0.5)
testthat::expect_equal(extract_grading("1.0_test"), 1.0)
testthat::expect_equal(extract_grading("1.5_test"), 1.5)
testthat::expect_equal(extract_grading("2.5_test-with-long-label"), 2.5)
}
\seealso{
Other helpers: 
\code{\link{map_burngrading}()},
\code{\link{reactable_burngrading}()}
}
\concept{helpers}
